Allocated  theme into separate folder. Screen view isolated into page 'ScreenView'.
